+++ CHANGELOG.md
@@ -23,6 +23,7 @@
 #### :nail_care: Polish
 
 - Configuration fields `bs-dependencies`, `bs-dev-dependencies` and `bsc-flags` are now deprecated in favor of `dependencies`, `dev-dependencies` and `compiler-flags`. https://github.com/rescript-lang/rescript/pull/7658
+- Better error message if platform binaries package is not found. https://github.com/rescript-lang/rescript/pull/7698
 
 #### :house: Internal